+++ example/app/urls.py
@@ -35,4 +35,6 @@ urls = {
     'ws/<user_id>/': UserWebsocket,
     'send/<connection_id>/': send_message_to_websocket_api,
     'bg-tasks/': run_background_tasks_api,
+    'custom-response/': custom_response_class_api,
+    'image/': ImageAPI,
 }