+++ docs/docs/release_notes.md
@@ -1,3 +1,14 @@
+### 3.1.3
+- Add `Image` base class 
+- Add `size` to `File` base class
+- Improve the way of loading `configs` in `single-file` structure
+- Improve `background_tasks.py`, `generate_ws_connection_id()`
+- `bpython` removed from being the default python shell
+- Improve `load_middlewares()` error handling 
+- Print `configs` on the `run`
+- Add `requirements.txt` for development 
+- Update `roadmap.jpg`, `README.md`
+ 
 ### 3.1.2
 - Add new methods to `BackgroundTask`
   - `every_seconds()`